dt1 editor: broken long files
